requirements.txt
Django==5.0.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
requests==2.31.0
celery==5.4.0
django-celery-beat==2.6.0
redis==5.0.4
django-cors-headers==4.3.1
coverage==7.5.0
flake8==7.0.0